attempt to fix CI issue by moving AWS S3 V3 test into reader.js
